+++ tests/test_mc_plot.py
@@ -70,9 +70,13 @@ def plot_testing():
     var2.addVarStat(stattype='orderstatTI', statkwargs={'p': 0.6, 'c': 0.50, 'bound': '2-sided'})
     var2.addVarStat(stattype='mean')
 
-    mc_plot(var2, highlight_cases=None)              # mc_plot_2d_line
+    mc_plot(var2, highlight_cases=None)               # mc_plot_2d_line
     mc_plot(var1, var2, highlight_cases=[0, 1])       # mc_plot_2d_line
     mc_plot(var1, var2, var3, highlight_cases=[0, ])  # mc_plot_3d_line
+
+    m = np.eye(3)
+    var4 = MCOutVar(name='testm', vals=[1*m, 2*m, 0*m, -1*m, -2*m])
+    mc_plot(var4, highlight_cases=[])                 # mc_plot_3d_line
 
     mc_plot_cov_corr(np.array([[2, 0.1111, np.nan], [-0.19, -1, np.nan], [np.nan, np.nan, np.nan]]),
                               ['Test1', 'Test2', 'Test3'])
@@ -84,6 +88,7 @@ def plot_testing():
                                     refval=0.5, conf=0.95)
     mc_plot_integration_error(mcinvars['randint2'], volume=1, dimension=1,
                               refval=0.5, conf=0.95)
+    plt.show()
 
 
 if __name__ == '__main__':